requirements.txt
langchain-core 
langchain-cli 
langchain-community 
langchain-openai 
opensearch-py 
fastapi 
uvicorn 
grpcio 
grpcio-tools
python-keycloak
python-dotenv